Add config option to ignore untracked git files
The `parse_git_dirty` function currently does a git status which includes all untracked files. Not only doe some repos have a lot of auto generated files, but in large code bases doing a git status without filtering untracked files can take a long time. This change adds an option so it is possible to change the behavior of the command.