Disable Subversion by default.
This mirrors a choice made in upstream fish (as of version 3.1.2).
/usr/bin/svn on macOS is now a stub and is very slow by default
causing significant delays in the fish prompt coming up. Given
Subversion is not a particularly wide-spread VCS these days and
given the upstream choice to disable it, this disables Subversion
in the prompt by default but allows it to be easily re-enabled.